--- a/.github/workflows/integrations.yml
+++ b/.github/workflows/integrations.yml
@@ -67,7 +67,7 @@ jobs:
           repo-token: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Compile Contract
-        run: cargo build -p mpc-contract --target wasm32-unknown-unknown --release
+        run: cargo build -p mpc-contract --target wasm32-unknown-unknown --release --target-dir target/seperate_wasm
 
       - name: Test
         run: cargo test -p mpc-recovery-integration-tests mpc --jobs 1 -- --test-threads 1