--- a/.github/workflows/multichain-contract.yml
+++ b/.github/workflows/multichain-contract.yml
@@ -31,7 +31,7 @@ jobs:
       - uses: Swatinem/rust-cache@v1
 
       - name: Compile Contract
-        run: cargo build -p mpc-contract --target wasm32-unknown-unknown --release
+        run: cargo build -p mpc-contract --target wasm32-unknown-unknown --release --target-dir target/seperate_wasm
 
       - name: Test Contract
         run: cd contract && cargo test --target x86_64-unknown-linux-gnu --release