--- a/integration-tests/src/multichain/mod.rs
+++ b/integration-tests/src/multichain/mod.rs
@@ -2,6 +2,7 @@ pub mod containers;
 pub mod local;
 
 use crate::env::containers::DockerClient;
+use crate::mpc::TARGET_CONTRACT_DIR;
 use crate::{initialize_lake_indexer, LakeIndexerCtx};
 use mpc_contract::primitives::CandidateInfo;
 use mpc_recovery_node::gcp::GcpService;
@@ -153,7 +154,7 @@ pub async fn setup(docker_client: &DockerClient) -> anyhow::Result<Context<'_>> 
 
     let mpc_contract = worker
         .dev_deploy(&std::fs::read(
-            "../target/wasm32-unknown-unknown/release/mpc_contract.wasm",
+            format!("{}/wasm32-unknown-unknown/release/mpc_contract.wasm", TARGET_CONTRACT_DIR),
         )?)
         .await?;
     tracing::info!(contract_id = %mpc_contract.id(), "deployed mpc contract");