@DavidM-D 2024-03-19
On the following benchmark

```rust
#[test(tokio::test)]
async fn overhead_bench() -> anyhow::Result<()> {
    with_multichain_nodes(MultichainConfig::default(), |ctx| {
        Box::pin(async move {
            panic!();
        })
    })
    .await
}
```
Before:
```
cargo test multichain::overhead_bench -- --exact --test-threads 1  619.91s user 25.71s system 602% cpu 1:47.19 total
```

After
```
cargo test multichain::overhead_bench -- --exact --test-threads 1  1.43s user 0.78s system 14% cpu 15.621 total
```